+++ cmd/niac/root.go
@@ -8,7 +8,7 @@ import (
 )
 
 var (
-	version = "v2.1.0"
+	version = "v2.8.1"
 	commit  = "dev"
 	date    = "unknown"
 )